Support the case were the Jenkins build result is null.
Normally this will never be the case because the plugin
is a Recorder which can only be run as a post-build step.
However, with use of the "Any Build Step Plugin", we
can be added as a normal build step and are exposed
to the fact that a running build can have a build result
of null (if no build step has set it yet).